change to uint256 gas: sum:  17135581  avg:  114237.20666666667  all:  [50266,105620,71420,71420,118441,103297

+++ POC/src/HashTowerWithHashList.sol
@@ -9,8 +9,8 @@ Poseidon2 constant P2 = Poseidon2(0x1111111122222222333333334444444400000002);
 
 // capacity = W * (W**0 + W**1 + ... + W**(H - 1)) = W * (W**H - 1) / (W - 1)
 // 4 * (4^0 + 4^1 + 4^2 + ... 4^19) = 1_466_015_503_700
-uint8 constant H = 20;
-uint8 constant W = 4;
+uint256 constant H = 20;
+uint256 constant W = 4;
 uint256 constant CAPACITY = 1_466_015_503_700; // https://github.com/ethereum/solidity/issues/13724
 
 uint256 constant SNARK_SCALAR_FIELD =
@@ -19,7 +19,7 @@ uint256 constant SNARK_SCALAR_FIELD =
 
 contract HashTowerWithHashList {
     //HashTowerWithHashListH20W4Verifier private immutable verifier = new HashTowerWithHashListH20W4Verifier();
-    uint64 private _count;
+    uint256 private _count;
     uint256[H] private D;
     uint256[H] private DD;
 
@@ -28,27 +28,27 @@ contract HashTowerWithHashList {
     function add(uint256 toAdd) public {
         require(toAdd < SNARK_SCALAR_FIELD, "HashTowerWithHashList: toAdd must be < SNARK_SCALAR_FIELD");
 
-        uint64 count = _count;
+        uint256 count = _count;
         require(count < CAPACITY, "HashTowerWithHashList: full");
 
-        uint64 z;
-        uint64 fl;
-        uint8 ll; // TODO: or just uint256 ?
-        uint8 lv;
-        uint64 W_pow_lv = 1; // W ** lv
+        uint256 z;
+        uint256 fl;
+        uint256 ll;
+        uint256 lv;
+        uint256 W_pow_lv = 1; // W ** lv
 
         // find the lowest level that has space
         while (true) {
             z += W_pow_lv;
             fl = (count < z) ? 0 : (count - z) / W_pow_lv + 1;
-            ll = uint8((fl == 0) ? 0 : (fl - 1) % W + 1);
+            ll = (fl == 0) ? 0 : (fl - 1) % W + 1;
             if (ll < W) break;
             lv++;
             W_pow_lv *= W;
         }
         // append and go downward
         uint256 v = (lv > 0) ? D[lv - 1] : toAdd;
-        emit Add(lv, fl, v);
+        emit Add(uint8(lv), uint64(fl), v);
         uint256 d = (ll == 0) ? v : P2.poseidon([D[lv], v]);
         uint256 dd = (fl == ll) ? d : P2.poseidon([DD[lv + 1], d]); // fl == ll means there is no level above
         uint256 prevDd;
@@ -63,14 +63,14 @@ contract HashTowerWithHashList {
             // the rest levels are all full
             fl = (count - z) / W_pow_lv + 1;
             v = (lv > 0) ? D[lv - 1] : toAdd;
-            emit Add(lv, fl, v);  // emit event
+            emit Add(uint8(lv), uint64(fl), v);
             d = v;
             dd = P2.poseidon([prevDd, d]);
         }
         _count = count + 1;
     }
 
-    function getCountAndDd() public view returns (uint64, uint256) {
+    function getCountAndDd() public view returns (uint256, uint256) {
         return (_count, DD[0]);
     }
     /*
